refactor(scan-dsl): extract node scan dsl
