+++ app/src/main/java/com/neo/speaktouch/utils/extension/AccessibilityNodeInfo.kt
@@ -185,3 +185,5 @@ fun AccessibilityNodeInfo.indexOfChild(
 
     error("Child not found")
 }
+
+val AccessibilityNodeInfo.lastIndex get() = childCount - 1
